Remove defaultOptions from ng-chat component and move unit tests to window spec file
