+++ src/spec/ng-chat.component.spec.ts
@@ -1272,44 +1272,6 @@ describe('NgChat', () => {
         expect(() => subject.initializeTheme()).toThrow(new Error(`Invalid theme configuration for ng-chat. "${subject.theme}" is not a valid theme value.`));
     });
 
-    it('Must return default chat options exercise', () => {
-        let chattingTo = new User();
-        let currentWindow = new Window(chattingTo, false, false);
-
-        let result = subject.defaultWindowOptions(currentWindow);
-
-        expect(result).not.toBeNull();
-        expect(result.length).toBeGreaterThanOrEqual(1);
-        expect(result[0].displayLabel).toBe("Add People");
-        expect(result[0].action).not.toBeNull();
-        expect(result[0].validateContext).not.toBeNull();
-
-        expect(result[0].validateContext(chattingTo)).toBeTruthy();
-        expect(result[0].validateContext(new Group([]))).toBeFalsy();
-    });
-
-    it('Must return empty chat options when participant is not an user', () => {
-        let chattingTo = new Group([]);
-        let currentWindow = new Window(chattingTo, false, false);
-
-        let result = subject.defaultWindowOptions(currentWindow);
-
-        expect(result).not.toBeNull();
-        expect(result.length).toBe(0);
-    });
-
-    it('Must return empty chat options when group adapter is not supplied', () => {
-        let chattingTo = new User();
-        let currentWindow = new Window(chattingTo, false, false);
-
-        subject.groupAdapter = null;
-
-        let result = subject.defaultWindowOptions(currentWindow);
-
-        expect(result).not.toBeNull();
-        expect(result.length).toBe(0);
-    });
-
     it('onOptionPromptCanceled invoked should clear selection state', () => {
         let mockedUser = new User();
         mockedUser.id = 999;
